Example of D-bus connection between server and client
The example implements server and client applications and allows to
communicate them via the system bus.
The server app creates a connection to the system bus, creates a D-bus
object, registers methods and signals and runs I/O event loop.
The client app creates a proxy object for the server object, subscribes to
the signals, invokes the method and receives and handles the signal.

Signed-off-by: Ilia Skochilov <iskochilov@ilbers.de>